requirements.txt
boto3==1.21.25
coverage==6.3.2
cryptography==36.0.2
Django==4.0.3
django-filter==21.1
django-multiselectfield==0.1.12
django-storages==1.12.3
djangorestframework==3.13.1
Pillow==9.0.1
psycopg2-binary==2.9.3
pytest==7.1.1
pytest-cov==3.0.0
pytest-django==4.5.2
python-dotenv==0.20.0
pytz==2022.1
gunicorn==20.1.0
